Declutter delegates & move to POC package
Addresses #4. Move all existing delegates to a POC package for temporary
holding area. Some of these will be promoted down the road as we
determine which are applicable.
