+++ src/main/java/org/folio/rest/delegate/poc/AbstractOkapiRequestRuntimeDelegate.java
@@ -1,9 +1,9 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import org.camunda.bpm.engine.RuntimeService;
 import org.springframework.beans.factory.annotation.Autowired;
 
-public abstract class AbstractRuntimeDelegate extends AbstractLoggingDelegate {
+public abstract class AbstractOkapiRequestRuntimeDelegate extends AbstractOkapiRequestDelegate {
 
   @Autowired
   protected RuntimeService runtimeService;
