+++ src/main/java/org/folio/rest/delegate/poc/AbstractRuntimeDelegate.java
@@ -1,9 +1,9 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import org.camunda.bpm.engine.RuntimeService;
 import org.springframework.beans.factory.annotation.Autowired;
 
-public abstract class AbstractOkapiRequestRuntimeDelegate extends AbstractOkapiRequestDelegate {
+public abstract class AbstractRuntimeDelegate extends AbstractLoggingDelegate {
 
   @Autowired
   protected RuntimeService runtimeService;
