+++ src/main/java/org/folio/rest/delegate/poc/CrUpdateLoanStatusDelegate.java
@@ -1,6 +1,7 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
+import org.camunda.bpm.engine.impl.util.json.JSONObject;
 import org.camunda.bpm.engine.variable.Variables;
 import org.camunda.bpm.engine.variable.value.ObjectValue;
 import org.camunda.spin.json.SpinJsonNode;
@@ -15,18 +16,7 @@ import org.springframework.stereotype.Service;
 import static org.camunda.spin.Spin.JSON;
 
 @Service
-public class GenericOkapiRequestDelegate extends AbstractRuntimeDelegate {
-
-  private static final String REQUEST_URL = "requestUrl";
-  private static final String REQUEST_METHOD = "requestMethod";
-  private static final String REQUEST_PAYLOAD = "requestPayload";
-  private static final String REQUEST_URI_VARIABLES = "requestUriVariables";
-
-  private static final String REQUEST_CONTENT_TYPE = "requestContentType";
-
-  private static final String RESPONSE_STATUS = "responseStatusName";
-  private static final String RESPONSE_BODY = "responseBodyName";
-  private static final String RESPONSE_HEADER = "responseHeaderName";
+public class CrUpdateLoanStatusDelegate extends AbstractRuntimeDelegate {
 
   @Value("${tenant.headerName:X-Okapi-Tenant}")
   private String tenantHeaderName;
@@ -34,11 +24,18 @@ public class GenericOkapiRequestDelegate extends AbstractRuntimeDelegate {
   @Autowired
   private OkapiRequestService okapiRequestService;
 
+  @Value("${okapi.location}")
+  private String OKAPI_LOCATION;
+
   @Override
   public void execute(DelegateExecution execution) throws Exception {
-    log.info("Executing Generic Okapi Request Delegate");
+    log.info("Executing Update Loan Status Delegate");
 
     String tenant = execution.getTenantId();
+    String loanId = execution.getProcessBusinessKey();
+    String userId = execution.getVariable("userId").toString();
+    String itemId = execution.getVariable("itemId").toString();
+    SpinJsonNode checkOutJson = JSON(execution.getVariable("checkOutJson").toString());
 
     String okapiToken = "";
     if (execution.getVariable("folioLogin") != null) {
@@ -46,16 +43,29 @@ public class GenericOkapiRequestDelegate extends AbstractRuntimeDelegate {
       okapiToken = folioLogin.getxOkapiToken();
     }
 
-    SpinJsonNode jsonNode = JSON(execution.getVariable("okapiRequest"));
+    String requestUrl = String.format("%s/circulation/loans/%s", OKAPI_LOCATION, loanId);
+    log.info("requestUrl: {}", requestUrl);
+    String requestMethod = "PUT";
+    String requestContentType = "application/json";
+    String responseStatusName = "";
+    String responseHeaderName = "";
+    String responseBodyName = "";
 
-    // TODO: Refactor to map directly to OkapiRequest
-    String requestUrl = jsonNode.prop(REQUEST_URL).stringValue();
-    String requestMethod = jsonNode.prop(REQUEST_METHOD).stringValue();
-    String requestContentType = jsonNode.prop(REQUEST_CONTENT_TYPE).stringValue();
-    String responseStatusName = jsonNode.prop(RESPONSE_STATUS).stringValue();
-    String responseHeaderName = jsonNode.prop(RESPONSE_HEADER).stringValue();
-    String responseBodyName = jsonNode.prop(RESPONSE_BODY).stringValue();
-    SpinJsonNode payload = jsonNode.prop(REQUEST_PAYLOAD);
+    JSONObject json = new JSONObject();
+    json.put("id", loanId);
+    json.put("userId", userId);
+    json.put("itemId", itemId);
+
+    JSONObject item = new JSONObject();
+    JSONObject itemStatus = new JSONObject();
+    itemStatus.put("name", "Lost");
+    item.put("status", itemStatus);
+
+    json.put("item", item);
+    json.put("loanDate", checkOutJson.prop("loanDate").stringValue());
+    json.put("action", "lost");
+
+    SpinJsonNode payload = JSON(json.toString());
 
     OkapiRequest okapiRequest = new OkapiRequest();
     okapiRequest.setTenant(tenant);
@@ -68,17 +78,16 @@ public class GenericOkapiRequestDelegate extends AbstractRuntimeDelegate {
     okapiRequest.setRequestPayload(payload);
     okapiRequest.setOkapiToken(okapiToken);
 
-    log.info("JSON: {}", jsonNode);
     log.info("payload: {}", payload);
 
     OkapiResponse okapiResponse = okapiRequestService.okapiRestCall(okapiRequest);
-    log.info("OKAPI RESPONSE: {}", okapiResponse);
+    log.info("OKAPI RESPONSE UPDATE LOAN STATUS: {}", okapiResponse);
 
     ObjectValue response = Variables.objectValue(okapiResponse)
       .serializationDataFormat("application/json")
       .create();
 
-    execution.setVariable("okapiResponse", response);
+    execution.setVariable("okapiResponseUpdateLoanStatus", response);
 
   }
