+++ src/main/java/org/folio/rest/delegate/poc/FolioLoginDelegate.java
@@ -1,4 +1,4 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import static org.camunda.spin.Spin.JSON;
