+++ src/main/java/org/folio/rest/delegate/poc/ParseOkapiResponseDelegate.java
@@ -1,4 +1,4 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.spin.json.SpinJsonNode;
