+++ src/main/java/org/folio/rest/delegate/poc/ResetLoginTokenDelegate.java
@@ -1,4 +1,4 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.folio.rest.model.FolioLogin;
