+++ src/main/java/org/folio/rest/delegate/poc/TestProcessDelegate.java
@@ -1,4 +1,4 @@
-package org.folio.rest.delegate;
+package org.folio.rest.delegate.poc;
 
 import java.nio.charset.Charset;
